refactor: change FipePrice according to new data modeling
