+++ src/main/java/com/tcc5/car_price_compare/domain/vehicle/Vehicle.java
@@ -1,7 +1,6 @@
 package com.tcc5.car_price_compare.domain.vehicle;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.tcc5.car_price_compare.domain.price.FipePrice;
 import com.tcc5.car_price_compare.domain.shared.TimestampedEntity;
 import com.tcc5.car_price_compare.domain.user.features.Notification;
 import jakarta.persistence.*;
@@ -42,7 +41,6 @@ public class Vehicle extends TimestampedEntity {
     @JsonIgnore
     private List<Notification> notifications = new ArrayList<>();
 
-    @OneToOne
-    @JoinColumn(name = "price")
-    private FipePrice fipePrice;
+    @OneToMany(mappedBy = "vehicle", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    private List<FipePrice> fipePrices = new ArrayList<>();
 }
